Change 'server online' notification from webscoket to REST API response

+++ host/index.js
@@ -7,41 +7,45 @@ let serverProcess = null;
 const ws = new WebSocket("ws://192.168.3.61:1987");
 
 fastify.get("/start-server", async (request, reply) => {
-  serverProcess = spawn(
-    "cmd.exe",
-    ["/c", path.resolve("./enigmatica6/start-server.bat")],
-    { stdio: "pipe" }
-  );
+  try {
+    await new Promise((resolve, reject) => {
+      serverProcess = spawn(
+        "cmd.exe",
+        ["/c", path.resolve("./enigmatica6/start-server.bat")],
+        { stdio: "pipe" }
+      );
 
-  serverProcess.on("error", (error) => {
-    reply.code(500).send(`Failed to start server: ${error.message}`);
-    console.error(`Failed to start server: ${error.message}`);
-  });
+      serverProcess.on("error", (error) => {
+        reject(new Error(`Failed to start server: ${error.message}`));
+        console.error(`Failed to start server: ${error.message}`);
+      });
 
-  serverProcess.stdout.on(
-    "data",
-    (data) =>
-      data
-        .toString()
-        .includes(
-          "Loading of 'lightnetwork' for world minecraft:overworld finished"
-        ) && ws.send("server-on")
-  );
+      serverProcess.stdout.on("data", (data) => {
+        if (
+          data
+            .toString()
+            .includes(
+              "Loading of 'lightnetwork' for world minecraft:overworld finished"
+            )
+        )
+          resolve();
 
-  serverProcess.on("exit", () => ws.send("server-off"));
+        if (data.toString().includes("Press any key to continue"))
+          serverProcess.kill();
+      });
 
-  reply.code(200).send("Server starting");
+      serverProcess.on("exit", () => ws.send("server-off"));
+    });
+
+    reply.code(200).send("Server started");
+  } catch (error) {
+    reply.code(500).send(error.message);
+  }
 });
 
 fastify.get("/stop-server", async (request, reply) => {
   serverProcess.stdin.write("/stop");
   serverProcess.stdin.end();
-  serverProcess.stdout.on(
-    "data",
-    (data) =>
-      data.toString().includes("Press any key to continue") &&
-      serverProcess.kill()
-  );
 
   // Hacky way to see if I'm using the PC
   await exec(
@@ -56,17 +60,9 @@ fastify.get("/stop-server", async (request, reply) => {
 const start = async () => {
   try {
     await fastify.listen({ port: 3000, host: "192.168.3.61" });
-  } catch (err) {
-    fastify.log.error(err);
+  } catch (error) {
+    fastify.log.error(error);
     process.exit(1);
   }
 };
 start();
-
-
-
-
-
-
-
-
